+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- JSONSchema validate:
+  - Add format checkers
+
 ## [0.3.0] - 2022-11-08
 
 ### Added